% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Simulate data from an N-of-1 trial with washout}
\usage{
simulate(
  n_trts,
  n_blocks,
  period_length,
  washout_length,
  sampling_timestep,
  sd_b,
  sd_p,
  sd_o,
  Ejs,
  gammas,
  taus,
  eta,
  custom.Bt = NULL,
  Xj_inits = NULL,
  Z_init = NULL
)
}
\arguments{
\item{n_trts}{The number of treatments being tested.}

\item{n_blocks}{The number of blocks, where each block consists of a sequence of treatments.}

\item{period_length}{The length of each treatment period.}

\item{washout_length}{The length of the washout period between treatments, during which no treatment is administered.}

\item{sampling_timestep}{The time interval at which observations are sampled.}

\item{sd_b}{Standard deviation of the baseline process.}

\item{sd_p}{Standard deviation of the process noise.}

\item{sd_o}{Standard deviation of the observation noise.}

\item{Ejs}{A vector of treatment effect sizes for each treatment (Ej).}

\item{gammas}{A vector of decay constants for the decay of treatment effects when the treatment is not being applied (gamma).}

\item{taus}{A vector of decay constants for the decay of treatment effects when the treatment is being applied (tau).}

\item{eta}{Time constant for the decay of the effect of the total treatment on the outcome (Zt).}

\item{custom.Bt}{Custom baseline function.}

\item{Xj_inits}{Custom initial values for treatment effect functions.}

\item{Z_init}{Custom initial value for the outcome process (Zt).}
}
\value{
a list with three main components: "data", "functions", and "params"
\itemize{
\item \code{data}: A tibble (data frame) containing the simulated data for the N-of-1 trial with columns:
\itemize{
\item code{time}: Observation time points.
\item code{period}: Period number corresponding to each time point.
\item code{treatment}: Treatment administered at each time point.
\item code{Bt}: Baseline values at each time point.
\item code{sum_Xjt}: Total treatment effect at each time point.
\item code{Zt}: Simulated outcome without error at each time point.
\item code{Yt}: Simulated outcome with error at each time point.
\item code{Tj}: Binary indicators for each treatment at each time point (one column per treatment).
\item code{Xj}: Treatment effect for each treatment at each time point (one column per treatment).
}
\item \code{functions} A list of functions used during the simulation:
\itemize{
\item \code{Bt}: Baseline function
\item \code{Tt}: Function mapping time points to treatments.
\item \code{Xjt}: List of treatment effect functions for each treatment.
\item \code{Zt}: Simulated outcome function without error.
-\code{params}: A list of parameter settings:
\item \code{trial}: List of trial parameters, including number of treatments (n_trts), number of blocks (n_blocks), period length (period_length), and washout length (washout_length).
\item \code{process}: List of process parameters, such as time steps (timesteps), sampling time step (sampling_timestep), and standard deviations for baseline (sd_b), process noise (sd_p), and observation noise (sd_o).
\item \code{treatment}: List of treatment effect parameters, including treatment effect sizes (Ejs), decay constants for treatment effects when treatment is not being applied (gammas), decay constants for treatment effects when treatment is being applied (taus), and the time constant for the decay of the effect of the total treatment on the outcome (eta).
}
}
}
\description{
This function simulates data from an N-of-1 trial, which is a single-subject
clinical trial in which the subject serves as their own control. The function
generates a dataset that includes time, period, treatment, baseline,
treatment effects, simulated outcome without error, and simulated outcome
with error.
}
